Robot work.
Added the loop block, only one block remains.
Changed how the json is formatted to reduce size.

+++ Arduino/Reader.h
@@ -5,7 +5,7 @@
 #include "ArduinoJson.h"
 #include "Sphero.h"
 
-int runCode(Sphero sphero, char* serializedCode);
+uint8_t runCode(Sphero sphero, char* serializedCode);
 uint8_t runBlock(Sphero sphero, JsonObject block);
 
 #endif
